+++ NEWS.md
@@ -1,3 +1,7 @@
+# version 0.9-6
+
+* use `s2::s2_dwithin_matrix` in `st_is_within_distance`; #1367
+
 # version 0.9-5
 
 * Only when package `s2` >= 1.0.1 is available: support for spherical geometry operators (predicates, transformers, measures, nearest point/feature) for geographic coordinates in package `s2` is now by default switched off, and can be switched on by `sf_use_s2(TRUE)`; see https://www.r-spatial.org/r/2020/06/17/s2.html and vignette sf7. It is planned to be switched on by default in sf 1.0-0.